moved some patches
