+++ Plugin/LegacyPatches/UnityEngine/InputPatch.cs
@@ -146,128 +146,6 @@ public class InputPatch
     }
     */
 
-    [HarmonyPatch(typeof(InputOrig), nameof(InputOrig.GetKeyInt))]
-    private class GetKeyInt
-    {
-        private static Exception Cleanup(MethodBase original, Exception ex)
-        {
-            return PatcherHelper.CleanupIgnoreException(original, ex);
-        }
-
-        private static bool Prefix(object key, ref bool __result)
-        {
-            if (reverseInvokerFactory.GetReverseInvoker().Invoking)
-                return true;
-            var env = virtualEnvironmentFactory.GetVirtualEnv();
-            if (!env.RunVirtualEnvironment) return true;
-            __result = env.InputState.KeyboardState.Keys.Contains((int)(KeyCode)key);
-            return false;
-        }
-    }
-
-    // above gets called from GetKey
-
-    [HarmonyPatch(typeof(InputOrig), nameof(InputOrig.GetKeyString))]
-    private class GetKeyString
-    {
-        private static Exception Cleanup(MethodBase original, Exception ex)
-        {
-            return PatcherHelper.CleanupIgnoreException(original, ex);
-        }
-
-        private static bool Prefix( /*string name, ref bool __result*/)
-        {
-            if (reverseInvokerFactory.GetReverseInvoker().Invoking)
-                return true;
-            var env = virtualEnvironmentFactory.GetVirtualEnv();
-            return !env.RunVirtualEnvironment;
-            // TODO
-        }
-    }
-
-    // above gets called from GetKey
-
-    [HarmonyPatch(typeof(InputOrig), nameof(InputOrig.GetKeyUpString))]
-    private class GetKeyUpString
-    {
-        private static Exception Cleanup(MethodBase original, Exception ex)
-        {
-            return PatcherHelper.CleanupIgnoreException(original, ex);
-        }
-
-        private static bool Prefix( /*string name, ref bool __result*/)
-        {
-            if (reverseInvokerFactory.GetReverseInvoker().Invoking)
-                return true;
-            var env = virtualEnvironmentFactory.GetVirtualEnv();
-            return !env.RunVirtualEnvironment;
-            // TODO
-        }
-    }
-
-    // above gets called from GetKeyUp
-
-    [HarmonyPatch(typeof(InputOrig), nameof(InputOrig.GetKeyUpInt))]
-    private class GetKeyUpInt
-    {
-        private static Exception Cleanup(MethodBase original, Exception ex)
-        {
-            return PatcherHelper.CleanupIgnoreException(original, ex);
-        }
-
-        private static bool Prefix(object key, ref bool __result)
-        {
-            if (reverseInvokerFactory.GetReverseInvoker().Invoking)
-                return true;
-            var env = virtualEnvironmentFactory.GetVirtualEnv();
-            if (!env.RunVirtualEnvironment) return true;
-            __result = env.InputState.KeyboardState.KeysUp.Contains((int)(KeyCode)key);
-            return false;
-        }
-    }
-
-    // above gets called from GetKeyUp
-
-    [HarmonyPatch(typeof(InputOrig), nameof(InputOrig.GetKeyDownString))]
-    private class GetKeyDownString
-    {
-        private static Exception Cleanup(MethodBase original, Exception ex)
-        {
-            return PatcherHelper.CleanupIgnoreException(original, ex);
-        }
-
-        private static bool Prefix( /*string name*/)
-        {
-            if (reverseInvokerFactory.GetReverseInvoker().Invoking)
-                return true;
-            var env = virtualEnvironmentFactory.GetVirtualEnv();
-            return !env.RunVirtualEnvironment;
-            // TODO
-        }
-    }
-
-    // above gets called from GetKeyDown
-
-    [HarmonyPatch(typeof(InputOrig), nameof(InputOrig.GetKeyDownInt))]
-    private class GetKeyDownInt
-    {
-        private static Exception Cleanup(MethodBase original, Exception ex)
-        {
-            return PatcherHelper.CleanupIgnoreException(original, ex);
-        }
-
-        private static bool Prefix(object key, ref bool __result)
-        {
-            if (reverseInvokerFactory.GetReverseInvoker().Invoking)
-                return true;
-            var env = virtualEnvironmentFactory.GetVirtualEnv();
-            if (!env.RunVirtualEnvironment) return true;
-            __result = env.InputState.KeyboardState.KeysDown.Contains((int)(KeyCode)key);
-            return false;
-        }
-    }
-
-    // above gets called from GetKeyDown
 
     [HarmonyPatch(typeof(InputOrig), nameof(InputOrig.GetAxis))]
     private class GetAxis
